+++ conf/script/build-system/meson/setup/fetch_gcc_version.py
@@ -3,7 +3,7 @@
 import colorama
 
 from data_model import Compiler
-from fetch_compiler_version import cli_fetch_gnu_compiler_version
+from cli_fetch_gnu_compiler_version import cli_fetch_gnu_compiler_version
 
 # Run as a script
 if __name__ == '__main__':
